patches from denper (mms) and gazstone (online media info). thanks guys!

+++ source/wiimc.cpp
@@ -77,7 +77,7 @@ void CheckSleepTimer()
 	if(sleepStart == 0)
 		return;
 
-	if(diff_sec(sleepStart, gettime()) > WiiSettings.sleepTimer*60)
+	if(diff_sec(sleepStart, gettime()) > (u32)(WiiSettings.sleepTimer*60))
 	{
 		ShutdownRequested = 1;
 		controlledbygui = 2;
